requirements.txt
Flask==3.1.2
protobuf==6.33.0
pycryptodome==3.23.0
Requests==2.32.5